pyproject.toml
[project]
name = "inventory-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.41.0",
    "pandas>=2.3.3",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.50.0",
]
